Remove the Alias/Host and Theme abstractions
This project started as a Wordpress Mu wannabe. Support for multi-blog
serving and user themes was half-baked in, but, let's face it, this is
never going to develop into a full-blown Blogger. And if it were to, i'd
rewrite it from scratch with other, better suited, technologies. So, it
was time to simplify the model and get rid of some layers that were
never used properly.